KDTree periodic boundaries issue
These changes fix an issue with the handling of the periodic boundary conditions I have been having recently. Using the old looping method, some particles which were outside the box were not being correctly moved inside the opposite boundary, meaning that cKDTree was throwing an error and crashing the code.

Furthermore, the cKDTree function is now identical to the KDTree function, so I have update this also. Though it may prove wise to not do this for better backwards compatability with scipy<1.6 as discussed in the function's scipy documenatation.